Add the asyncio executor: real await-concurrency for ASGI (Phase 2)
Zero-config ASGI now serves through one executor thread per event loop
(m0_wsgi.asgi_executor) running the bridge's persistent asyncio loop,
fed through the unchanged OffloadPool: the loop parks each request and
submits its slot as one datagram, loop.add_reader turns slots into
tasks, and completions answer through put_response/complete -- the
"any producer holding the pool address" property the offload design
already guaranteed. Requests overlap wherever the application awaits:
eight concurrent 1.5s awaits complete in 1.51s on one loop with zero
threads, where the buffered bridge takes 12s. The banner says
asgi-loop; use_asgi_executor decides (explicit --blocking-threads N
keeps the Phase-1 buffered pool as the escape hatch), and --threads N
composes with one executor per loop.

The executor path never builds a WSGI environ: spawn_asgi crosses
method, path and protocol as str, the query as bytes, the headers as a
ready list of lowercase (bytes, bytes) pairs, and the body as bytes --
all stolen slots of one args tuple through one PyObject_CallObject.
That replaced an environ + Python-side scope re-transform that did the
header work twice, and the RSS guard now reads 20KB over 10k executor
requests. _serve_one_buffered awaits the app directly (no inner task,
no asyncio.wait arbiter -- the removed scaffolding was a measured
seven percent of hello-world throughput) with the streaming watchdog
moved into send(). Exactly one lifespan runs per event loop: fallback
handlers are built with lifespan=False and the executor's own app owns
startup/shutdown. The executor picks uvloop for its own loop where
installed, stdlib asyncio otherwise.

poe bench-asgi is the standing gate against uvicorn (single process,
stdlib-client caveat recorded): the mixed slow/fast fast-request p99
passes (2.87ms vs 3.27ms) and hello-world throughput stands at
0.88-0.94x, with the remainder located -- the WSGI pool row clears
uvicorn through the same offload machinery, so it is the one executor
thread plus the per-request datagram round trip -- and fix paths
recorded in docs/WSGI_PERFORMANCE.md. smoke-asgi's phases now pin the
executor shape (asgi-loop banner, no x-thread, eight-way overlap, RSS,
watchdog, SIGTERM drain) plus the escape hatch and the WSGI zero-config
pool; test_cli covers use_asgi_executor and the flipped ASGI default.

Also fixes a latent Int/Int32 mismatch in the fork's dead
kevent_register, surfaced when the new import graph elaborated it.

+++ packages/m0-http/lightbug_http/c/kqueue.mojo
@@ -128,7 +128,7 @@ def kevent_register(kq: FileDescriptor, changes: Span[kevent_t, ...]) raises:
     for i in range(n):
         cl[i] = changes[i]
 
-    var result = _kevent(kq.value, cl, c_int(n), None, c_int(0), None)
+    var result = _kevent(Int32(kq.value), cl, c_int(n), None, c_int(0), None)
     cl.unsafe_free()
 
     if result == -1:
